refactor: folder structure for modal
