+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { RequiresAuth } from "./requiresAuth/RequiresAuth";
 export { Header } from "./header/Header";
-export { ThreadModal } from "./modal/ThreadModal";
+export { ThreadModal } from "./modal/threadModal/ThreadModal";
 export { FollowBar } from "./followBar/FollowBar";
 export { PostCard } from "./card/PostCard";
 export { QuestionCard } from "./card/QuestionCard";
